Refactor helper and customer logic
